requirements.txt
numpy>=1.26
torch>=2.0.1
gin-config>=0.5.0
